refactor: Dao 중복 코드 리펙터링
