+++ src/main/java/web/dao/ChessGameDao.java
@@ -1,38 +1,17 @@
 package web.dao;
 
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
 public class ChessGameDao {
-    public Connection getConnection() {
-        Connection con = null;
-        String server = "localhost:13306";
-        String database = "chess_db";
-        String option = "?useSSL=false&serverTimezone=UTC";
-        String userName = "root";
-        String password = "root";
-
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        } catch (ClassNotFoundException e) {
-            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
-            e.printStackTrace();
-        }
-
-        try {
-            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
-            System.out.println("정상적으로 연결되었습니다.");
-        } catch (SQLException e) {
-            System.err.println("연결 오류:" + e.getMessage());
-            e.printStackTrace();
-        }
-
-        return con;
-    }
+    private static final DatabaseConnection databaseConnection = new DatabaseConnection();
 
     public void addCommand(String command, int roomId) throws SQLException {
-        try (Connection con = getConnection()) {
+        try (Connection con = databaseConnection.getConnection()) {
             String query = "INSERT INTO chessGame(command, room_id) VALUES (?, ?)";
             PreparedStatement pstmt = con.prepareStatement(query);
             pstmt.setString(1, command);
@@ -44,7 +23,7 @@ public class ChessGameDao {
     public List<String> findByRoomId(int roomId) throws SQLException {
         List<String> moves = new ArrayList<>();
 
-        try (Connection con = getConnection()) {
+        try (Connection con = databaseConnection.getConnection()) {
             String query = "SELECT command FROM chessGame WHERE room_id = ? ORDER BY command_id";
             PreparedStatement pstmt = con.prepareStatement(query);
             pstmt.setInt(1, roomId);
