+++ src/main/java/web/dao/RoomDao.java
@@ -1,41 +1,19 @@
 package web.dao;
 
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
 public class RoomDao {
+    private static final DatabaseConnection databaseConnection = new DatabaseConnection();
     private static final int NO_ROOM_NUMBER = 0;
-
-    public Connection getConnection() {
-        Connection con = null;
-        String server = "localhost:13306";
-        String database = "chess_db";
-        String option = "?useSSL=false&serverTimezone=UTC";
-        String userName = "root";
-        String password = "root";
-
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        } catch (ClassNotFoundException e) {
-            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
-            e.printStackTrace();
-        }
-
-        try {
-            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
-            System.out.println("정상적으로 연결되었습니다.");
-        } catch (SQLException e) {
-            System.err.println("연결 오류:" + e.getMessage());
-            e.printStackTrace();
-        }
-
-        return con;
-    }
 
     public int newRoomId() throws SQLException {
         int roomId = NO_ROOM_NUMBER;
-        try (Connection con = getConnection()) {
+        try (Connection con = databaseConnection.getConnection()) {
             String query = "INSERT INTO chessRoom() VALUES()";
             PreparedStatement pstmt = con.prepareStatement(query);
             pstmt.executeUpdate();
@@ -53,7 +31,7 @@ public class RoomDao {
 
     public List<Integer> findRoomIds() throws SQLException {
         List<Integer> roomIds = new ArrayList<>();
-        try (Connection con = getConnection()) {
+        try (Connection con = databaseConnection.getConnection()) {
             String query = "SELECT room_id FROM chessRoom";
             PreparedStatement pstmt = con.prepareStatement(query);
             ResultSet rs = pstmt.executeQuery();
