+++ src/test/java/web/dao/RoomDaoTest.java
@@ -3,10 +3,6 @@ package web.dao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import web.dao.RoomDao;
-
-import java.sql.Connection;
-import java.sql.SQLException;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -16,14 +12,6 @@ public class RoomDaoTest {
     @BeforeEach
     public void setup() {
         roomDao = new RoomDao();
-    }
-
-    @Test
-    @DisplayName("db 연결 확인")
-    public void connection() throws SQLException {
-        try (Connection con = roomDao.getConnection()) {
-            assertNotNull(con);
-        }
     }
 
     @Test
